+++ logconfig/exceptions.py
@@ -1,0 +1,11 @@
+"""Exceptions used by this module.
+"""
+
+__all__ = (
+    'LogConfigException',
+)
+
+
+class LogConfigException(Exception):
+    """Base exception for logconfig module."""
+    pass