+++ CHANGES.rst
@@ -2,6 +2,10 @@ Changelog
 =========
 
 
+- Rename project to ``logconfig``.
+- Rename ``ConfiglogException`` to ``LogConfigException``. (**breaking change**)
+
+
 v0.3.1 (2014-12-23)
 -------------------
 
@@ -21,7 +25,7 @@ v0.3.0 (2014-12-23)
 v0.2.0 (2014-12-23)
 -------------------
 
-- Renamed project to ``configlog``. (**breaking change**)
+- Rename project to ``configlog``. (**breaking change**)
 - Rename ``ConfigException`` to ``ConfigLogException``. (**breaking change**)